requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.35.0
easyocr>=1.7.0
pytesseract>=0.3.10
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0

# NLP and Text Processing
spacy>=3.7.0
nltk>=3.8.0
sentencepiece>=0.1.99

# Explainability (Optional - custom visualizations are implemented)
# shap>=0.44.0  # Optional for advanced explainability
# lime>=0.2.0.1  # Optional for advanced explainability

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
python-dateutil>=2.8.2

